doc(protocol): Improve documentation on event types (#954)
Improves incomplete documentation on event types. The description largely
repeats itself so that information can be found in both places. The is one
slight difference: On the event field, we do not mention specific event types
since SDKs are not supposed to send them. For more details, one can still look
up the enumeration docs.