feat: Log extra metrics for internal project ids (#975)
Co-authored-by: Jan Michael Auer <mail@jauer.org>

We are observing a large amount of broken traces due to missing transactions. Together with getsentry/sentry#25188 this should make it more clear where they are dropped. We still need a PR to make it clearer how many transactions the SDK sent, because we were previously counting envelopes that include sessions.

#skip-changelog