fix(server): Do not wait on request data after error (#1006)
Streams in futures 0.1 are allowed to resume after emitting an error. It
appears that the actix Payload type locks up when being polled after
erroring since it does not signal EOF in all error cases. To avoid this,
we remove all use of the raw Payload in favor of a fully fused
SharedPayload.