fix(schema): Explicitly declare reprocessing context (#1009)
Reprocessing adds context data to an event that helps in finding those
events based on their previous issue ID. We have observed that, since
this is event data coming out of processing, it can happen that
datascrubbing (invoked at the end of process_event) removes this vital
information, or more specifically, replaces the original_primary_hash
with a SHA1 hash due to a PII rule. This SHA1 hash is passed to Snuba
which fails to parse it as UUID. This caused the Snuba consumer to
crash, hotfix at getsentry/snuba#1896

To fix this I define the reprocessing context explicitly, set pii=false,
and make sure it is omitted from the exported JSON schema. This commit
also contains changes to make it possible to omit an entire enum variant
from the schema. Previously it was only possible to omit fields.

In the store normalizer I make sure the reprocessing context is
removed. NormalizeProcessor is not invoked as part of renormalization in
Sentry, only in Relay.

While building this I also noticed that the process_trace_context was
never wired up correctly, so I fixed that too.