fix(store): Use random Kafka partition key for session aggregates (#1022)
Session aggregates come with a session_id of NIL. This causes all
aggregated sessions to be routed to the same Kafka partition. To avoid
this, we assign a random partition key. Without a key, Kafka would fall
back to hashing the message, which has the same effect but is less
explicit in application code.

Note that session updates are fully additive and commutative. It's
therefore better to distribute them evenly, rather than pinning to some
natural key.

We temporarily apply the same fallback for metrics.