ref(server): Remove noisy payload parsing errors (#1032)
Removes errors when event payloads or session payloads cannot be parsed.
These were reported as "failed to extract event" and "failed to store
session", respectively.

Such errors are completely expected when invalid payloads are sent into
Relay. They were originally introduced to debug broken SDKs, but in the
way they are being reported they are neither actionable to the Relay
development team, nor anyone running Relay onpremise.

Eventually, we will bring this back in form of a metric including the
SDK version, and potentially a mechanism to sample invalid payloads. For
now, these errors are not actionable, and drown real issues in Sentry.