feat(config): Support running Relay without config directory (#1055)
Previously, the existence of some valid config inside some config directory (.relay by default) was necessary for Relay to launch.

This change allows the user to perform basic configuration of Relay solely through the environment variables or commandline arguments.

The names of the variables are not changed from what was already used to override file-based config. There is a new variable RELAY_MODE and an analogous commandline argument for setting the mode. Explicitly providing a non-existing config directory via CLI argument continues to raise an error.

Co-authored-by: Markus Unterwaditzer <markus-honeypot@unterwaditzer.net>