ref(metrics): Remove the aggregator queue (#1067)
To simplify metrics aggregation, remove the Aggregator's internal queue.
Instead, poll the entire list of buckets every time. By default, this will
remove about 10% of the full list of buckets in every flush cycle.

As opposed to a dedicated queue, filtering the buckets might be slightly slower.
However, the bulk of the work is then performed outside of the aggregator. In
turn, we use less memory on the second queue structure.