feat(client-reports): Protocol support for client reports
This adds support for client reports from the SDKs. This feature is primarily
used to allow client SDKs to emit outcomes.

The way this is currently implemented is that we add a new envelope item
type called client_report which contains these outcomes. The envelope item
type is intentionally kept flexible for future expansion as we already discussed
emitting other internal debug information upstream which is unrelated to
outcomes (eg: number of breadcrumbs discarded and configuration errors).

Because of how SDKs are currently implemented we now reserve internal
as data category for such metric items. Relay however will never emit
such a data category as we do not want to communicate actual rate limits
for such envelope items.

From the relay perspective any outcome reason that an SDK emits is
acceptable and upstreamed. It's up to the snuba consumer to whitelist
these outcomes to avoid writing bad data into the kafka topic. This
way SDKs can send more outcomes even if relays have not been updated.