fix(metrics): Accumulate session metrics on session start time (#1105)
Fix session metrics extraction such that session metrics are always
emitted with the session start time.

We treat the number of started sessions in each time bucket as the total
session count for that time bucket, and compute the crash rate, for
example, by dividing by this total. It is therefore necessary to emit
all consequent session updates into the same time bucket. This was
always the intention but must have slipped during the initial
implementation.