feat(core): Add option to control allocation sizes
Add a config option to control the maximum allowed allocation size. On
failed allocation the thread will crash, and actix will maybe restart
it, but clearly not crash the entire process.

This was done in response to us finding a problem with actix's handling
of content encoding, where it would decompress a <1MB zipbomb into
multiple gigabytes allocated in a contiguous buffer.

There really should not be a reason to ever allocate 1GB,
at least not on a whim and for a single event.

One concern is regexes: those take a lot of memory and we're fine with
that, because we assume that them taking more space would shave off CPU
load. I think we'd be fine with a regex taking 1GB of RAM under certain
circumstances, but right now they don't so it's fine.