fix(server): Finish decoding compressed requests correctly (#1123)
It is required to finish decoders to flush remaining buffers and validate
checksums at the end of the payload. In this PR, we ensure this by replacing
Decoder::take with Decoder::finish. The take method is now fully internal
and only used in the streaming DecodingPayload. All other users must call
finish instead.