feat(metrics): Another statsd metric to measure bucket duplication [INGEST-421]
Add a set metric that measures the number of unique bucket keys observed at a point/interval in time.

This requires us to implement some way of hashing bucket keys such that statsd can consume them. BucketKey already implements Hash to be used in hashmaps. We cannot however use the std hasher:

* docs for DefaultHasher explicitly state that the hashes may change across rust releases (i guess this would not be too much of a blocker for our purpose, but it's annoying to keep in mind)

* SipHasher is deprecated (but could probably be used)

we already have crc32fast in our dependency tree so let's depend on it
explicitly and just use that

there's still one caveat in that the impls of Hash may call different methods depending on cpu architecture (as stated in https://docs.rs/deterministic-hash/1.0.1/deterministic_hash/), but I think we can live with that for now.