ref(metrics): Ability to use static strs in metrics
This is usually marginally better than using String everywhere (e.g. for
events it would not be very useful, we'd need more dynamic string
interning there), but in the particular case where we extract metrics
from sessions, a lot of our tags and metric names come from static
strings.

The insert_metrics microbenchmark suite for the metrics aggregator
registers an improvement of 5% across the board (consistently for all
benchmarks, so I would think this is not just noise). But the benchmark
is meaningless, we really need to measure session extraction.