feat: Capture minidumps for hard crashes (#1127)
Uses sentry-native with the breakpad backend to offer an optional
crash-handler feature for Linux and macOS. When enabled, it'll capture
hard crashes and write minidumps with stack traces of all threads into a
database folder on the file system. On next restart, those crashes will
be sent to Sentry for debugging. The DSN and release are inherited from
the main Sentry SDK.