fix(pii): Ignore `replacement_chunks` when they aren't used (#1180)
With ReplaceBehavior::Value, we only want to replace a string value.
This means we don't need the replacement chunks since
`insert_replacement_chunks` correctly adds the replacement and the
`process_text` function is useless because we don't need it (besides,
processing an empty string doesn't perform any action).

Now that the `replacement_chunks` are only needed in
ReplaceBehavior::Groups, we can move the processing and assertion we do
on them to that branch of the `match`. This solves the issue of
asserting on them when we don't use them.