ref(metrics): Add custom and more builtin units [INGEST-939] (#1217)
Adds a number of grouped builtin units and a "custom unit" variant.

Units within the same group are considered convertible. Each group has a
default canonical unit, for instance `byte` for information units. In
the future, Relay can streamline incoming metrics by converting to the
canonical unit within the respective group. The currently defined
builtin units are:

- Duration — nanosecond / microsecond / millisecond / second / minute /
  hour / day / week
- Information — bit / byte / kibibyte / mebibyte / gibibyte / tebibyte /
  pebibyte / exbibyte
- Fraction — percent / ratio
- None — none, "" (used for counts)

For forward compatibility, clients can send custom unit names of up to
15 ASCII characters. Beyond ASCII lowercasing, these units do not count
as convertible. An example for such a unit would be `frame` to track the
frame rate of a user interface or game.