feat(general): Map Windows version from raw_description to name (#1219)
Parsing the Windows version from os.raw_description would previously
return the raw version in the form of <major>.<minor>.<build-number>.
This PR attempts to map the build number to well known version names
like "Vista" or "11" instead. In case the build number is not in the
list of known builds, we fall back to the raw version.

Fixes #1201.