fix(metrics): Remove/reject nul-bytes from metric strings [ingest-1204]
We are observing release names which are UTF-16 strings reinterpreted as
UTF-8. This is causing issues downstream and does not line up with the
releases that end up in the event payload, which for some reason are
totally correct.

In order to counteract this, we strip NUL-bytes from release values.

In other string types (tag keys, metrics), there are "fewer excuses" for
not getting those strings right, so we just drop them.