ref(constants): Remove the unused "internal" data category (#1245)
DataCategory::Internal was originally introduced to handle client
reports, but did not end up used. Instead, client reports like many
other items do are not rate limited and as such do not have any data
category.

This patch removes the unused variant and the only reference to it from
a code comment.