feat(metrics): Add client sdk as tag to session metrics (#1248)
Adds an additional `sdk` tag to session metrics, sourced from the
`sentry_client` header. This header usually contains the SDK name and
version in the format `name/version`, for example `raven-node/2.6.3`. In
practice, this tag is not going to increase dimensionality since the
client version correlates with the `release` tag.