fix(alloc): Remove global allocator
Our custom allocator in relay that we use to prevent oom contains ub.

GlobalAlloc is documented such that alloc/dealloc must not unwind, which means no panics
https://doc.rust-lang.org/std/alloc/trait.GlobalAlloc.html#safety

I probably overlooked this because the documentation was not on the unsafe methods themselves

on discord i learned that this documentation exists because the __rust_alloc symbol is annotated with the nounwind flag in llvm. this permits llvm to produce optimizations that can cause memory unsafety should we choose to unwind

I'm also convinced that the goal of "shoot down the thread" requires
unwinding. Since all known bugs are fixed, let's rather remove this
mitigation than risk running into UB.