ref: Use Duration to compute breakdowns and span durations (#1260)
Instead of using raw f64, this patch switches to using Duration for
computing span durations and breakdowns. This allows to pass the values
around without the additional knowledge in which unit they are.
Previously, this was described only in comments.

This change also simplifies the computation algorithm. There are three
repeated changes:

- Remove redundant code. Especially in `time_spent`, there were
  redundant allocations when collecting spans.
- When converting durations to float milliseconds, do not drop the
  sub-millisecond part.
- Use interval clamping rather than a series of comparisons.