feat(server): Retry project configs asynchronously (#1263)
Introduces version 3 of the endpoint protocol for project configs. It is
a superset version 2 that adds a list of `pending` keys. Keys in that
list are computed asynchronously by the upstream and should be retried
after a short while.

The version 2 endpoint blocked the response of the entire batch request
on every requested project config. Especially for longer chains of
Relays, a single slow config could therefore fail or timeout an entire
batch. Under load or in extraordinary circumstances, this had the
potential to accelerate the degradation of ingestion.

The version 3 implementation is twofold:

1. The upstream project source puts every project key from the `pending`
   list back into the queue. It will be retried on the next fetch
   cycle(s), until a config can be retrieved or the project exceeds its
   retry limits.
2. The endpoint now supports both versions, 2 and 3. In version two, it
   continues to await the project state response and never populates
   `pending`. In version 3, it only responds with a config if the cached
   state is up-to-date (including soft outdated). Otherwise, version 3
   always responds with `pending`.

An example for the response schema is:

    {
      "configs": {
        "50f0f59961bce997c872b45942c57e9f": { /* ... */ },
        "8f19b1b4d628e4ef0a3014c70379c180": { /* ... */ }
      },
      "pending": ["e1eef6aef6b20fcdb8190d703ab6a56b"]
    }