build: Use Centos 7 as base image for the linux build (#1279)
Moves docker builds to a container based on CentOS 7, which is currently
the oldest long-term supported Linux still under maintenance along with
RHEL 7. This makes Relay compatible with glibc 2.17 or newer. Apart from
CentOS and RHEL, this change also makes Relay compatible with the
current AWS lambda runtime.

The main docker container published to the DockerHub registry continues
to include processing and crash-handling, which comes with an additional
dependency to curl. To make this work under CentOS, there was a fix
necessary in the relay-crash native build to ensure the static library
gets installed into the correct directory.