feat: Add DataCategory constant for processed-transactions
While Relay won't be emitting this DataCategory for now, we need
a distinct category for processed transactions as usage tracking and
billing in general is oriented around data categories that come from
relay.

I'm not confident this is the best name but I thought this would be
a good way to get a discussion started.