ref(project_upstream): Add counter for completed projects
We do not have any indication of completed requests, especially when
they are unsuccessful due to timeouts they just silently disappear.
This adds a counter to keep track of completed requests and their
status.

It also fixes the attempts histogram to include the number of attempts
when the project state is given up on.