ref(metrics): Fewer metrics extracted for release health (#1316)
Since #1275, we extract user IDs into one or more of init, ok, errored,
crashed, abnormal buckets. The latter three are used by the product, but
init and ok can be consolidated into a single tag, because the product
does not query them explicitily as of getsentry/sentry#34858 and
getsentry/sentry#34957.

In order to reduce the number of buckets stored in clickhouse,
consolidate these tags.

Also, stop collecting session.duration for non-healthy sessions.