ref(dynamic_sampling): Simplify rule engine and callsite in envelope actor
We don't need two separate functions for matching events and traces for
as long as we're fine exposing the FieldValueProvider trait.

Also unify trace and event sampling in relay-server. This allows us to
make adjustments to the sample rate in one place.

Unifying it within relay-sampling turns out to provide little value
because trace sampling needs access to a different sampling config than
event sampling.