feat(protocol): Add transaction_info to events (#1330)
Implements `transaction_info` containing the `source` field, which
indicates how the event's transaction name was obtained on the client.
This allows us to infer the quality of the transaction name and apply
server-side cleanup. Additionally, this adds the
`event.transaction_source` metric to measure how often each source
occurs in practice.