fix(metrics): Consistent transaction.status (#1333)
In store normalization, we set contexts.trace.status to unknown in case
it is missing from the trace context. In metrics extraction, which runs
before normalization, we simply ignored a missing status. Align.

This does not increase dimensionality, since the bucket ..., status=none
will now move to bucket ..., status=unknown. There is already a static
string for unknown in the indexer, so this will not increase indexer
storage.

Future Work:
To reduce code duplication, carve out the parts of
store normalization that are needed for metrics extraction and put it
something like a PreMetricsProcessor that runs before extraction.