Fix wording: downstream -> upstream
Upstream relays are relays that are closer to the Sentry backend
(sentry.io). Downstream relays are relays that are closer to clients or
SDKs.

Events go from clients to sentry.io, which means they are forwarded
upstream. Metric extraction happens in the first relay configured to
extract metrics, and the following relays in chain (upstream relays)
must not extract metrics again. The word used in the comment is
"downstream", which according to the flow above, it's not possible to
happen.