feat(metrics): Extract transaction metrics in external relays (#1344)
Currently, transaction metrics are only extracted in processing relays
(the function call to extract metrics is surrounded by an
if_processing!). Even if we extract metrics in external relays, we don't
want to extract metrics multiple times from a transaction, causing
metric (and billing) duplication.

The implemented approach in this PR is to extract metrics in the very
first relay in the chain, mark the transaction as extracted, and not
extract metrics from that transaction in subsequent relays. The simplest
way to accomplish that is using the existing metrics_extracted header in
items that had metrics extracted. The approach and the header are reused
from the metric extraction from sessions, there's nothing new here.

Implementation-wise, removing that if_processing triggers a lot of
changes to make current processing-only features non-processing too. On
the other hand, we store a new flag transaction_metrics_extracted on
ProcessEnvelopeState to make it available to the whole state processing
function and transfer it to the outgoing event. Ideally, process_state
is refactored not to throw every piece of data we need to the envelope
state object, but that's out of the scope of this PR.