light normalize test_fixtures before store processing
warning: the snapshot update is the result of a breaking change
introduced in the light normalization, where the user agent is always
parsed no matter what the config looks like. previously, this was
disabled by default and you had to explicitly enable user agent
normalization. we accept this for now, and there will be work scheduled
in the future to address this.