fix(http): Explain why we responded with 429
fix https://github.com/getsentry/sentry/issues/37103

Oftentimes we are not in control of the rate limit response, but if we
are, we can provide a more helpful error message.

The context here is that a customer was confused about whether the
network errors in Chrome's dev console would have any impact on their
application. While we can't influence the color of what is shown in the
dev console, we can provide documentation links in request/response
details. This will still be quite hidden but better than nothing.

Other ideas were to stop emitting 429s, but we could only do that for
browser, mobile and other "client" SDKs. For server-side SDKs we _need_
them to see a 429 so they stop sending data. And even then that change
would come with high risk of multiplying inbound traffic by a large
factor.