ref(server): Split EnvelopeManager file into smaller parts (#1388)
The file containing EnvelopeManager defined two other major components:
First, the EnvelopeProcessor which made up most of the file's contents
and contains all synchronous processing code calling out to the
normalization libraries. Second, the EnvelopeContext which comprises
relevant information to carry along with Envelopes.

This PR separates all of these components into their own files so that
changes to the EnvelopeManager's internals can be reviewed better. The
precise changes are:

- Move EnvelopeProcessor along with all its messages into a dedicated
  file
- Move EnvelopeContext into `utils::with_outcome`, which is what it is
  intended for.
- Clean up duplication in EncodeEnvelope. This processing function can
  subsequently be removed.

There is no change to control flow or behavior of any involved
components.