ref(server): Remove overall envelope handling timeout (#1398)
Removes the overall timeout for Envelope buffering and handling,
controlled by the `cache.envelope_expiry` option. There is a range of
timeouts for project configs and upstream requests that individually
control these processes. The overall buffer is controlled by
`cache.envelope_buffer_size`.

A reasonable overall time limit is hard to set. In many cases, either
the limit is too short to bridge real downtimes, or it's too long for
the buffer size. In practice, we've been relying on the count-based
limit exclusively. With proper disk-based caching, the limit could be
set to a product-limit such as 30 days.

The option remains in the config and is now marked as unused. It is
likely that some form of time-based expiry is re-introduced once
disk-based caching is implemented.