feat(actix): Generic Addr type
This makes `Addr` generic including the return value, just for the
Healthcheck actor for now.

The `HealthcheckMessage` has been renamed to `HealthcheckEnvelope` and
still needs to be a manual list of all the messages the actor
implements.  The conversion from messages to this is now a little more
involved that an `impl From`, since this also encodes the return type
so you need to type this out with the `ActorMessage` trait impl.

Sending to the actors is now generic, and the actors main loop also
remains reasonable clean to implement.