ref(server): Localize outcome emission in EnvelopeContext (#1406)
This change makes `EnvelopeContext` responsible for tracking the
lifetime of an envelope and emitting outcomes. In a follow-up, the goal
is to remove the long-running handler future in `EnvelopeManager`, and
instead pass envelopes through separate queues the ingestion stages.
Most importantly, this requires to have a central place that is bound to
the envelope and emits outcomes and metrics when envelopes are dropped
for any reason.

## EnvelopeContext

Before this change, `EnvelopeContext` was a data object containing all
information to create outcomes. It is primarily used in three places:

- The endpoint handler future, where it was shared across all future
  callbacks to emit outcomes when envelopes are dropped.
- The envelope manager's handler future, in exactly the same way.
- The envelope processor, where it was stored on the processing state
  and used to emit outcomes for inbound filtering and rate limiting.

The envelope context is now passed alongside the envelope through Relay
and also moves through message handlers. This can ensure that outcomes
are guaranteed when the envelope is dropped. It can be dropped in two
ways:

- Explicitly by calling `accept()` or `reject(outcome)`.
- Implicitly by dropping, for instance if sending a message fails
  unexpectely. This represents a bug and will record an `"internal"`
  outcome.

Calling `accept()` consumes the envelope context. Temporarily,
`reject()` does not yet consume, which will be changed in a follow-up.

## Control Flow

The flow of envelopes remains unchanged, with the envelope context now
following the envelope through every step:

(in the endpoint)

1. Load the request and construct an envelope. Drop empty envelopes.
2. Send it to the project cache for the fast-path. This may drop the
   envelope for cached rate limits or disabled projects. Outcomes are
   emitted in the project cache. Updates the envelope context with
   project information if available.
3. Send it to the envelope manager for queueing. If queueing fails, an
   outcome is emitted by the envelope manager.
4. End the request.

(in envelope manager)

1. Split envelopes into event-related items, metrics items, and
   everything else. This will create duplicate envelope contexts, since
   from here on there are multiple independent envelopes that get
   queued.
2. Send the envelope through the project cache again, now ensuring an
   up-to-date project state. Same logic as in the endpoint.
3. Send the envelope to the processor. All rate limits and invalid
   transactions are logged directly by the processor. An updated
   envelope and envelope context are sent back to the Envelope Manager.
4. Submit the envelope to either Kafka, Upstream, or the internal
   capture map (only in capture mode). For simplicity, the envelope
   context remains in the waiting future outside.
5. Explicitly drop the envelope context either by accepting or rejecting
   based on the result of the previous step.