ref(server): Add a message to capture envelopes in capture mode (#1409)
Relay has a capture mode used for testing. In this mode it does not send
envelopes to Kafka or the upstream, but rather keeps them in an internal
map. There is a hidden endpoint with which a tester can extract the
captured envelopes and inspect them for errors.

Until now, envelopes were all handled and captured by the envelope
manager. The final goal is to remove the long-running handler future in
EnvelopeManager, which requires to capture rejections from other places
than the envelope manager.

This change introduces a message to capture envelopes or rejections from
any place in the code base. In subsequent refactors, individual error
handlers can send their errors into capture mode using this message.