ref(redis): keep the configuration in one place
Since there is already configuration for few different types of the
connection pools and redis clients, also most of the settings are
repeatable - it is time to move all those config options if one place.

Introducing `RedisConfigOptions` which will be a single place to gather
all the configuration options, with sensible defaults.