fix(server): Treat expired projects as absent (#1404)
Replace Project::state() and Project::state_clone() with
Project::get_expiry_state(), a function that returns both the Expiry of
a project state and the state itself.

By making this the only public interface to get the state of a project,
external clients won't be able to access expired state information,
meaning it won't be possible to apply outdated rate limits etc. on the
fast path.