feat(redis): Make connection pool configurable [INGEST-1557] (#1418)
* feat(redis): Make connections pool configurable [INGEST-1557]

Currently the maximum number of connections managed by the pool is
hardcoded and set to the magic number of `24` which works, but we must
make sure we have a posibility to finetune this parameter.

This change adds `max_connections` option into the `Cluster` config,
keeps the existing `Single` config server without changes for bakwards
compatibility and also introduces the `SingleOpts` variant to configure single
server with additional options (with e.g. `max_connections`).

The default number of redis connetions in the pool is kept to `24`.

This should allow us to keep existing configs without changes and still
give the posibility for finetunning connections in the future.

* ref(redis): keep the configuration in one place

Since there is already configuration for few different types of the
connection pools and redis clients, also most of the settings are
repeatable - it is time to move all those config options if one place.

Introducing `RedisConfigOptions` which will be a single place to gather
all the configuration options, with sensible defaults.