fix(normalization): Light normalize logentry
Error message filters look at the `formatted` and `message` entries of
an event, in that order. If `formatted` is not provided in the payload,
relay builds it from `message` and `params`, [see
docs](https://develop.sentry.dev/sdk/event-payloads/types/#typedef-LogEntry).
That build, aka logentry normalization, doesn't happen before filters
are applied. As a result, events with a `formatted` string that should
be filtered out are not, due to the string not being built yet.