feat: Add support for Kafka clusters sharding (#1454)
This allows to configured many different Kafka clusters per topic. The
basic idea is to provide the number of shards and the lower bound for
the range with the Kafka config name and the topic name on that cluster

```
 metrics:
     shards: 65000
     mapping:
         0:
             name: "ingest-metrics-1"
             config: "metrics_1"
         25000:
             name: "ingest-metrics-2"
             config: "metrics_2"
         45000:
             name: "ingest-metrics-3"
             config: "metrics_3"
```

This will also allow to configure one Kafka clusters and use different
topics names on it.