feat(protocol): Track transaction name changes (#1466)
Introduces `propagations` and `changes` that provides more insight into how a
transaction name was changed during the lifetime of a transaction span. Since
the transaction name is propagated to downstream SDKs as part of the Dynamic
Sampling Context (DSC), this is a relevant quality metric for sampling traces by
transaction name.

All traces where propagations occur before the final transaction name is set
may not be sampled consistently if there are rules matching on the transaction
name.