ref(actix): Port EnvelopeProcessor to tokio (#1470)
Removes the actix SyncArbiter in favor of a plain tokio service that
uses a semaphore around spawn_blocking to delegate CPU-intensive work.
The service is started in the main runtime since it does not exhibit a
large footprint. Tokio runtimes have a larger number of blocking threads
than required by the EnvelopeProcessor.

Tokio ensures that there are always threads to run blocking tasks. If
one of the tasks panics, drop handlers are still run, the panic handler
picks up the panic (reporting it to Sentry), and tokio starts a new
thread.