ref: Use simple modulo #shards to get the shard number
As a folllowup for #1454 we decided that we do not have to hash the org
id to get the sahrd number and simple modulo #shard will be sufficient.

As part of this small refactoring the `unwrap` was also removed and
Result is returned instead, to propagate the error to the caller if it
happens.