ref(service): Move from_registry to interface types (#1472)
When actors implemented `SystemService` directly, we used to obtain instances
via `<Actor>::from_registry`. Since new addresses bind to an interface rather
than service implementations, the registry getters should also be on the
interface class instead. This has the positive effect that there are no
references to the service classes in the code base outside of regular startup.